Fixes bz#5802. Additional error handling for fdebugrpm.
frysk-core/frysk/bindir/ChangeLog:
2008-02-28  Teresa Thomas  <tthomas@redhat.com>

	* fdebugrpm.sh: Show error message if fdebuginfo install
	or arguments missing.

+++ frysk-core/frysk/bindir/fdebugrpm.sh
@@ -42,32 +42,46 @@
 # Requires pid(s) as argument
 
 # Run fdebuginfo and get the name of missing debuginfo packages
-                                                                    
-export packages=`@bindir@/fdebuginfo "$*" | grep "\-\-\-" |   
-                 cut -d ' ' -f 1 | 
-                 sort | uniq | 
-                 grep '^/'| 
-                 xargs rpm -qf --qf '%{SOURCERPM}\n' | 
-                 sort | uniq | 
-                 sed -e 's/-/-debuginfo-/' |
-                 sed -e 's/.src.rpm//g'`
 
-if [ -n "$packages" ] 
+# Display error message if no args provided
+if [ $# -lt 1 ]
 then
-    # Display missing packages
-    echo ""
-    echo "Missing Debuginfo package(s)"
-    echo "============================"
-    echo "$packages"
-    echo ""
+    echo "No arguments provided."
+    exit
+fi
 
-    # Install on user request
-    read -p "Do you wish to install the above packages? [y/n]: " ch
-    if [ "$ch" = "y" ]  
+# Check if fdebuginfo is installed.
+if ! test -f "@bindir@/fdebuginfo"
+then
+    echo "fdebugrpm error: Frysk not installed."
+    
+else                                                               
+    export packages=`@bindir@/fdebuginfo $* | grep "\-\-\-" |   
+                     cut -d ' ' -f 1 | 
+                     sort | uniq | 
+                     grep '^/'| 
+                     xargs rpm -qf --qf '%{SOURCERPM}\n' | 
+                     sort | uniq | 
+                     sed -e 's/-/-debuginfo-/' |
+                     sed -e 's/.src.rpm//g'`
+
+    if [ -n "$packages" ] 
     then
-       sudo yum install --enablerepo=*-debuginfo $packages
-    fi
+        # Display missing packages
+        echo ""
+        echo "Missing Debuginfo package(s)"
+        echo "============================"
+        echo "$packages"
+        echo ""
 
-else
-   echo "No missing debuginfo packages"
+        # Install on user request
+        read -p "Do you wish to install the above packages? [y/n]: " ch
+        if [ "$ch" = "y" ]  
+        then
+            sudo yum install --enablerepo=*-debuginfo $packages
+        fi
+
+    else
+        echo "No missing debuginfo packages"
+    fi
 fi
